+++ packages/plugin-image/web/src/toolbar/createToolbar.js
@@ -17,7 +17,7 @@ export default function createToolbar({
       relValue,
       uiSettings,
       isMobile,
-      imageEditorWixSettings: settings.imageEditorWixSettings,
+      settings,
     }),
     InsertButtons: createInsertButtons({ helpers, t }),
     name: 'image',