+++ packages/plugin-image/web/src/toolbar/inline-buttons.jsx
@@ -5,11 +5,12 @@ import { MediaReplaceIcon, ImageEditorIcon } from '../icons';
 
 const removeEmpty = list => list.filter(item => !!item);
 
-export default ({ t, anchorTarget, relValue, uiSettings, isMobile, imageEditorWixSettings }) => {
+export default ({ t, anchorTarget, relValue, uiSettings, isMobile, settings = {} }) => {
   const modalStyles = getModalStyles({ isMobile });
   const imageEditorStyles = getModalStyles({
     customStyles: { content: { maxWidth: '100%', background: 'transparent' } },
   });
+  const { imageEditorWixSettings, onImageEditorOpen } = settings;
   const imageEditorButton = imageEditorWixSettings
     ? {
         keyName: 'imageEditor',
@@ -19,6 +20,7 @@ export default ({ t, anchorTarget, relValue, uiSettings, isMobile, imageEditorWi
         modalStyles: imageEditorStyles,
         t,
         imageEditorWixSettings,
+        onImageEditorOpen,
         mobile: false,
         tooltipTextKey: 'ImageEditorButton_Tooltip',
         mapComponentDataToButtonProps: componentData => ({