+++ CHANGELOG.md
@@ -14,6 +14,9 @@
     Click to see more.
   </summary>
   
+  ### :rocket: New Feature
+  - `image`
+    - [#476](https://github.com/wix-incubator/rich-content/pull/476) `config.onImageEditorOpen` is invoked when media studio is launched
   ### :bug: Bug Fix
   - `gallery`
     - [#469](https://github.com/wix-incubator/rich-content/pull/469) styles reference